Remove old map keys in reconcile (#30)
Fixes #29. The reconcile implementation for map did not remove keys from
the document which were not in the incoming map. Modify the
implementation to remove the old keys.

Unfortunately this currently requires cloning the incoming and existing
keys. To fix this we would need to detach the lifetime of the iterator
returned by `ReconcileMap::entries` from the `ReconcileMap`. This might
in principle be possible but for now we wait until someone has a
performance problem.