Add Text::update
Problem: sometimes when modifying text you are not able to capture the
edits to the text field as they happen. For example, if the text is in a
file and you just get notified when the file has changed then you have
no way of knowing which text was inserted and deleted. The `Text` API
requires that you express all changes as `splice` calls, so the user is
forced to figure out how to turn the new value into a set of `splice`
calls, which can be tricky.

Solution: add `Text::update`, which performs an LCS diff to figure out a
minimal set of `splice` calls to perform internally.