Add a `rename` attribute
Problem: sometimes the field names in the Automerge document are not
valid Rust identifiers, or the name is just not what you want to use in
the Rust code but the derive macro always uses the Rust field names to
determine the field names in the document.

Solution: Add a `rename` attribute which is analogous to serdes rename
attribute to specify the name to use in the document.