correctly use keys in MapReconciler::put and SeqReconciler::set
Problem: the logic for MapReconciler::put and SeqReconciler::set was not
using the keys of incoming data to determine whether to create a new
object or reconcile into an existing one. This meant that reconciling a
type with an existing document where the type matched but the keys did
not would lead to surprising behavior when merging as new objects would
not be created as expected.

Solution: update MapReconciler::put and SeqReconciler::set to examine
the keys of incoming data and create new objects when the keys do not
match existing objects.